third_party/WebKit/LayoutTests/fast/css-grid-layout/non-named-grid-line-get-set.html

<!DOCTYPE HTML>
<html>
<head>
<link href="resources/grid.css" rel="stylesheet">
<style>
.gridItem {
    grid-column: 1;
    grid-row: 1;
    height: 22px;
    width: 77px;
}
.gridWithFixed {
    grid-definition-columns: "first" 10px;
    grid-definition-rows: "first" 15px;
}
.gridWithPercent {
    grid-definition-columns: 53% "last";
    grid-definition-rows: 27% "last";
}
.gridWithAuto {
    grid-definition-columns: "first" auto;
    grid-definition-rows: auto "last";
}
.gridWithMinMax {
    grid-definition-columns: "first" minmax(10%, 15px);
    grid-definition-rows: minmax(20px, 50%) "last";
}
.gridWithFixedMultiple {
    grid-definition-columns: "first" "nav" 10px "last";
    grid-definition-rows: "first" "nav" 15px "last";
}
.gridWithPercentageSameStringMultipleTimes {
    grid-definition-columns: "first" "nav" 10% "nav" 15% "last";
    grid-definition-rows: "first" "nav2" 25% "nav2" 75% "last";
}
</style>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<div class="gridWithFixed" id="gridWithFixedElement"></div>
<div class="gridWithPercent" id="gridWithPercentElement"></div>
<div class="gridWithAuto" id="gridWithAutoElement"></div>
<div class="gridWithAuto" id="gridWithAutoWithChildrenElement">
    <div class="gridItem"></div>
</div>
<div class="gridWithMinMax" id="gridWithMinMax"></div>
<div class="gridWithFixedMultiple" id="gridWithFixedMultiple"></div>
<div class="gridWithPercentageSameStringMultipleTimes" id="gridWithPercentageSameStringMultipleTimes"></div>

<script src="resources/grid-definitions-parsing-utils.js"></script>
<script>
    description('Test that setting and getting grid-definition-columns and grid-definition-rows works as expected');

    debug("Test getting grid-definition-columns and grid-definition-rows set through CSS");
    testGridDefinitionsValues(document.getElementById("gridWithFixedElement"), "first 10px", "first 15px");
    testGridDefinitionsValues(document.getElementById("gridWithPercentElement"), "53% last", "27% last");
    testGridDefinitionsValues(document.getElementById("gridWithAutoElement"), "first auto", "auto last");
    testGridDefinitionsValues(document.getElementById("gridWithMinMax"), "first minmax(10%, 15px)", "minmax(20px, 50%) last");
    testGridDefinitionsValues(document.getElementById("gridWithFixedMultiple"), "first nav 10px last", "first nav 15px last");
    testGridDefinitionsValues(document.getElementById("gridWithPercentageSameStringMultipleTimes"), "first nav 10% nav 15% last", "first nav2 25% nav2 75% last");

    debug("");
    debug("Test getting and setting grid-definition-columns and grid-definition-rows through JS");
    testNonGridDefinitionsSetJSValues("'first' 18px", "66px 'last'", "first 18px", "66px last", "first 18px", "66px last");
    testNonGridDefinitionsSetJSValues("'first' 55%", "40% 'last'", "first 55%", "40% last", "first 55%", "40% last");
    testNonGridDefinitionsSetJSValues("'first' auto", "auto 'last'", "first auto", "auto last", "first auto", "auto last");
    testNonGridDefinitionsSetJSValues("'first' min-content", "min-content 'last'", "first min-content", "min-content last", "first min-content", "min-content last");
    testNonGridDefinitionsSetJSValues("'first' max-content", "max-content 'last'", "first max-content", "max-content last", "first max-content", "max-content last");
    testNonGridDefinitionsSetJSValues("'first' minmax(55%, 45px)", "minmax(30px, 40%) 'last'", "first minmax(55%, 45px)", "minmax(30px, 40%) last", "first minmax(55%, 45px)", "minmax(30px, 40%) last");
    testNonGridDefinitionsSetJSValues("'first' minmax(22em, max-content)", "minmax(max-content, 5em) 'last'", "first minmax(220px, max-content)", "minmax(max-content, 50px) last", "first minmax(22em, max-content)", "minmax(max-content, 5em) last", "first minmax(220px, max-content)", "minmax(max-content, 50px) last", "minmax(max-content, 50px) last");
    testNonGridDefinitionsSetJSValues("'first' minmax(22em, min-content)", "minmax(min-content, 5em) 'last'", "first minmax(220px, min-content)", "minmax(min-content, 50px) last", "first minmax(22em, min-content)", "minmax(min-content, 5em) last", "first minmax(220px, min-content)", "minmax(min-content, 50px) last", "minmax(min-content, 50px) last");
    testNonGridDefinitionsSetJSValues("'first' minmax(min-content, max-content)", "minmax(max-content, min-content) 'last'", "first minmax(min-content, max-content)", "minmax(max-content, min-content) last", "first minmax(min-content, max-content)", "minmax(max-content, min-content) last");
    testNonGridDefinitionsSetJSValues("'first' 'nav' minmax(min-content, max-content) 'last'", "'first' 'nav' minmax(max-content, min-content) 'last'", "first nav minmax(min-content, max-content) last", "first nav minmax(max-content, min-content) last", "first nav minmax(min-content, max-content) last", "first nav minmax(max-content, min-content) last");
    testNonGridDefinitionsSetJSValues("'first' 'nav' minmax(min-content, max-content) 'nav' auto 'last'", "'first' 'nav2' minmax(max-content, min-content) 'nav2' minmax(10px, 15px) 'last'", "first nav minmax(min-content, max-content) nav auto last", "first nav2 minmax(max-content, min-content) nav2 minmax(10px, 15px) last", "first nav minmax(min-content, max-content) nav auto last", "first nav2 minmax(max-content, min-content) nav2 minmax(10px, 15px) last");;
    testNonGridDefinitionsSetJSValues("'foo' 'bar' auto 'foo' auto 'bar'", "'foo' 'bar' auto 'foo' auto 'bar'", "foo bar auto foo auto bar", "foo bar auto foo auto bar", "foo bar auto foo auto bar", "foo bar auto foo auto bar");

    debug("");
    debug("Test getting and setting invalid grid-definition-columns and grid-definition-rows through JS");
    element = document.createElement("div");
    document.body.appendChild(element);
    element.style.gridDefinitionColumns = "'foo'";
    element.style.gridDefinitionRows = "'bar";
    shouldBeEqualToString("getComputedStyle(element, '').getPropertyValue('grid-definition-columns')", "none");
    shouldBeEqualToString("getComputedStyle(element, '').getPropertyValue('grid-definition-rows')", "none");

    element = document.createElement("div");
    document.body.appendChild(element);
    element.style.gridDefinitionColumns = "'foo' 'bar'";
    element.style.gridDefinitionRows = "'bar' 'foo'";
    shouldBeEqualToString("getComputedStyle(element, '').getPropertyValue('grid-definition-columns')", "none");
    shouldBeEqualToString("getComputedStyle(element, '').getPropertyValue('grid-definition-rows')", "none");
</script>
</body>
</html>
